training_db/requirements.txt
amqp==2.5.2
apturl==0.5.2
asgiref==3.2.3
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
billiard==3.6.1.0
bleach==3.1.0
blessings==1.7
Brlapi==0.6.6
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
diff-match-patch==20181111
distro-info===0.18ubuntu0.18.04.1
Django==2.2.8
django-allauth==0.41.0
django-celery-beat==1.5.0
django-excel==0.0.10
django-import-export==2.0.2
django-jet==1.0.8
django-master==0.2.7
django-mysql==3.2.0
django-rest-auth==0.9.5
django-richtextfield==1.5.0
django-timezone-field==4.0
djangorestframework==3.11.0
entrypoints==0.3
et-xmlfile==1.0.1
gunicorn==20.0.4
httplib2==0.9.2
idna==2.8
importlib-metadata==1.3.0
inquirer==2.6.3
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jdcal==1.4.1
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
keyring==10.6.0
keyrings.alt==3.0
kombu==4.6.7
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lml==0.0.9
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupPy==1.14
MarkupSafe==1.0
mistune==0.8.4
more-itertools==8.0.2
mysqlclient==1.4.6
nbconvert==5.5.0
nbformat==4.4.0
netifaces==0.10.4
notebook==5.7.8
oauth==1.0.1
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.45.1
openpyxl==3.0.3
ordereddict==1.1
pandocfilters==1.4.2
parso==0.4.0
pbr==5.4.4
pexpect==4.2.1
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.0.0
psycopg2==2.8.4
ptyprocess==0.6.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.8
Pygments==2.4.2
pygobject==3.26.1
pymacaroons==0.13.0
PyMySQL==0.8.0
PyNaCl==1.1.2
pyRFC3339==1.0
pyrsistent==0.15.2
python-apt==1.6.5+ubuntu0.2
python-crontab==2.4.0
python-dateutil==2.8.1
python-debian==0.1.32
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
pyzmq==18.0.1
qtconsole==4.5.1
readchar==2.0.1
redis==3.3.11
reportlab==3.4.0
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
requests2==2.16.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.13.2
six==1.13.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
tablib==1.1.0
terminado==0.8.2
testpath==0.4.2
testresources==2.0.1
texttable==1.6.2
tornado==6.0.2
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.7
usb-creator==0.3.3
vine==1.3.0
virtualenv==16.6.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
xkit==0.0.0
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
zope.interface==4.3.2
